Fix offset discrepancy

+++ backend/src/main/java/paintcontraints/backend/Constraint.java
@@ -1,4 +1,4 @@
 package paintcontraints.backend;
 
-public record Constraint(ColorComponents color, Operation operation, int[] indexes, double offset) {
+public record Constraint(ColorComponents color, Operation operation, int[] indexes, int offset) {
 }
